pyproject.toml
[tool.poetry]
name = "project_repo_tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
pysnyk = "^0.6.3"
retry = "^0.9.2"

[tool.poetry.dev-dependencies]
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
